+++ wxc_sdk/telephony/prem_pstn/route_group.py
@@ -151,10 +151,9 @@ class RouteGroupApi(ApiChild, base='telephony/config/premisePstn/routeGroups'):
         """
         params = org_id and {'orgId': org_id} or None
         body = update.model_dump_json(include={'name': True,
-                                    'local_gateways': {'__all__': {'trunk_id', 'priority'}}})
+                                               'local_gateways': {'__all__': {'trunk_id', 'priority'}}})
         url = self.ep(rg_id)
-        data = self.post(url=url, params=params, data=body)
-        self.put(url=url, params=params, data=data)
+        self.put(url=url, params=params, data=body)
 
     def delete_route_group(self, rg_id: str, org_id: str = None):
         """