+++ docs/changes.rst
@@ -2,6 +2,7 @@ Release history
 ===============
 
 - feat: device configurations API :attr:`wxc_sdk.WebexSimpleApi.device_configurations`
+- fix: :meth:`wxc_sdk.telephony.prem_pstn.route_group.RouteGroupApi.update`: used POST instead of PUT
 
 1.16.1
 ------